requirements.txt
Flask==2.3.3
Flask_Injector==0.15.0
flask_sqlalchemy==3.1.1
injector==0.21.0
pandas==2.1.1
pyfunctional==1.4.3
Requests==2.31.0
mysqlclient==2.2.0
pymysql==1.1.0
